Create README.md
The ideea of the project is that the administrators will have to add features, components and games. The user will be able to use this components to have a virtual copy of their own pc, and everytime they enter in one of the routes that are showing the requirements of a game, they will have a message-box that is showing if their pc can run the game based on the component's features and game's features.
In the future I will also like to add a news object and an editor role who will have the permission to add/edit the news object.